# Conflicts:
#	MockGroundStation/main.cpp

+++ Shared/UtilsLIB/UtilsLib/Utils.hpp
@@ -7,7 +7,7 @@ namespace utils
 {
   bool compareTwoDoubles(const double& a, const double& b);
 
-  bool checkWithin(size_t real, size_t close, size_t diff);
+  bool checkWithin(int real, int close, int diff);
 
   bool checkWithinDouble(double real, double close, double diff);
 }
